+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 This list is not intended to be all-encompassing - it will document major and breaking API 
 changes with their rationale when appropriate. Given version `A.B.C.D`, breaking changes are to be expected in version number increments where changes in the `A` or `B` sections:
 
+### v5.24.0.0
+- **http4k-connect-*** - Upgrade dependencies including Kotlin to 2.0.20
+- **http4k-connect-ai-openai*** - [Breaking] tightened up types for completion requests. 
+
 ### v5.23.0.0
 - **http4k-connect-*** - Upgrade dependencies including Kotlin to 2.0.20
 - **http4k-connect-ai-azure*** - Add new operations.